[LIGO-47] PascaLIGO: Move `block` out of `case_block` alternatives
Problem: In multisig.ligo:48, the case alternative block is
interpreted as empty, while it is not. This happens because
PascaLIGO parser expects to find `fields "statements"` inside
a `case_block`. In practice, if a `case_block` contains a `block`,
statements are deeper in the hierarchy, and get ignored.

Solution: Move "block" option out of the "case_block"
(to _if_clause).