[LIGO-66] Drop support for function application highlighting
Problem: It's hard to find function applications with regexes in
CameLIGO: if the general rule is "identifier followed by a group of
identifiers or parenthesized expressions separated by spaces" (which is
already kinda complex), we need to account for cases like `if x then` or
`f1 record.f2 boom` (we don't need to highlight f2 in this case). While
it is possible in theory, in practice it turns out to result in one huge
unmaintainable regex.

Solution: Drop function application highlighting. Add highlighting for
constructors with a simple rule.