Reimplement ASTMap with more features
The old ASTMap was a pretty straightforward cache without any internal
optimisation logic.

The new one is able to manage loading more efficiently, e.g. bundle
multiple loads together. It exposes a more flexible interface that
allows the user to choose what trade-off they want to accept.

* Replace `reload` with `invalidate`, which only marks data as outdated,
  but does not refetch it immediately.
* Replace `fetch` with `fetchCurrent`, which is more or less the same,
  except for some edge-cases when it will not start a new loading action
  if one is already running.
* `fetchLatest` provides stronger freshness guarantees, while
  `fetchBundled` and `fetchCached` provide weaker ones.

The idea is that in the future UI actions that the user can reasonably
expect to take some time and update the most up-to-date results should
use `fetchLatest`, while for events that cause the file to be reload but
also require feedback (e.g. showing something during typing)
`fetchBundled` or even `fetchCached` might be used.