Use the fetchBundled in the text change handler
It is slightly better than the previous implementation, because it
guarantees that there will not be a loading action per each keypress, if
keypresses happen quickly.

It is slightly worse, because it does not guarantee that when the user
stops typing the diagnostics will be up to date. This can be fixed later but
scheduling a delayed final update of diagnostics one second after the
user stops typing.

Also:

* Downgrade stm-containers to build with currently pinned hackage.nix.