Fix DocumentSymbol
The function that collects document symbols in response to the
documentSymbol request was always returning an empty list.

* Fix the matchers in the visitor that collects declarations to expect
  the tree structure with NameDecls instead of Names.
* Make the fallback scope builder collect TypeDecls.
* Fix `addLocalScopes` to not replace TypeName with Name in TypeDecl.
* Add a test.

The problem currently is that we are using the LSP interface that
returns a flat list of symbols rather than a nested tree. We are
returning not just top-level declarations, but also local variables and
constants. This might make sense in the nested case, but as a flat list,
this is just a mess.