DocumentSymbol: Do not return variables and constants
While we are using the flat symbols interface, returning variables and
constants declared inside functions results in a mess.

There doesn’t seem to be an easy way to only filter top-level
declarations, so we just no longer return any variables or constants at
all.

This is not optimal, since there probably can be top-level constants and
we want to collect them too.

Ideally, we should just switch to the interface that allows us to return
a nested tree.