DocumentSymbol: Only collect top-level declarations
Previously we were recursively collecting declarations everywhere in the
document. While we are returning a flat list of results, this was a
problem, since it resulted in a mess.

Instead of traversing the tree recursively, only collect the
declarations from the top level of the contract.

Also, bring back the collection of variables and constatants (in case
they can be on the top level).