Rename: Avoid duplication of the results
Previously, “reanem symbol” would try to edit the name at the
declaration place twice. This was happening because the function that
build the list of edits to make uses `referencesOf` to get the ranges to
edit. `referenceOf` returns not only actual references, but the
declaration palce itself.

Previously we would explicilty add `definitionOf` to the list of places
to edit. Now we don’t do this, since it is already included in
`referencesOf`.