Upgrade lsp: Step 2 = make it build
After bumping the version of lsp (previously haskell-lsp) some changes
needed to be made to the code to make it work again.

The biggest changes are in the Main module, since the handlers are now
specified in a completely different way. Syntactically, instead of a
record we now give them as a list. They are also more type-safe now,
since instead of manually constructing and sending a response, we are
now given a callback that accepts exactly the right type of the
response, manages the ids, and sends everything back.

For this reason, a number of utility functions are no longer needed.

The code of lsp has also been cleaned up, so some of the weird utils
it used to provide (e.g. something related to logging) no longer exist.
Previously, we use multiple different logging facilities, now we use
just one.

Another major change is that lsp now exposes its own monad transformer
`LspT` (which is just a ReaderT with their specific context, such as
`LspFunctions`, which we previously had to carry around ourselves) and a
`MonadLsp` typeclass. Rather than manually implementing an instance for
our `RIO` I chose to build our RIO on top of `LspT`, just as their docs
recommend. This is suboptimal, since we now have to Readers in the
stack, but it was easier and also, hopefully, will be harder to break
with further upgrades.

Configuration handling is now more straightforward too. When you are in
`MonadLsp`, you can request it anytime with `getConfig` and there is
only one configuration parsing handler. This way, lsp will keep track of
all configuration changes for us.

(Side note: IIUC, the handling of configuration in `lsp` is
incorrect, as the server only sends _changes_, while it is trying to
parse it as a whole value each time.)

Lastly, I have removed all the async dispatching from the main loop. I
have a suspicion that lsp might be doing just that anyway, but I
couldn’t find any definite confirmations in the docs. If it doesn’t, it
will be pretty easy to add it back (and I will do it in such a way that
manual dispatching of individual requests will not be required).