Add an empty document save handler
The new `lsp` logs a message if it tries to call a non-existent (on our
side) handler.

Previously we had document save and document close handlers, which were
empty, but they got removed when migrating to `lsp`.

This commit adds an empty handler for document save just to silence the
message. For document close we want to eventually have a non-empty
handler, so I am not silencing it here.