Add support for prepareRename
The prepareRename request allows the server to tell the client whether
it is possible to perform the rename operation at the given location
before actually starting it.

Add support for it by doing half the work of the actual rename handler:
we only lookup the declaration under the cursor and return its location
if we found something or `Nothing` if we did not.