Switch to the simpler rename interface
We used to return renaming results using `documentChanges` rather than
changes. This did not work, and my theory is that the reason was that we
were always setting document version to Nothing.

Switch to returning results via simpler `changes`, which only require
specifying the URI and not the version.

We currently do not use versions in any meaningful way, so it makes no
difference. We might want to switch back once we start actually doing
something useful with versions.