Lower the VSCode version restrictions, bump the plugin version
Problem: VSCode version seemed too restrictive, so I tested it the
plugin on older versions, and it seems to work starting from v1.41.
Moreover, we have released v0.2 already but the version in the repo
still shows v0.1.

Solution: Lower the version restrictions on VSCode, bump the plugin
version, update the readme.