[LIGO-91] Revert to Stardard strategy
Problem: We found ourselves that we never use `Standard` strategy
for parsing and use `parseWithScopes'` instead which is not what
we need.

Solution: Remove `parseWithScopes'` function in favor of using
`Standard` instance of scope forest. With that we have full control
of all possible errors that ligo may produce, by now we ignore all
but expected ones so that it won't clutter user debug output window.