[LIGO-113] Add 'extractTypeAlias' code action
Problem: We want to support type extraction for ligo meaning, that
we want to be able to extract any type and provide an alias for it
in one click.

Solution: Implement type alias extraction capability. Note that
there're a couple of issues arised within this MR, specifically:
  - vscode is very picky in chosing whether `changes` or
`document_changes` in code action are applied and does nothing if
both of them is set.
  - Somewhat we pass trees to code actions without any information
about scope variables, meaning that we are currently unable to
generate fresh new ones when defining type alias.
  - `alanz/lsp` package is currently not able to send multiple
requests per code action, meaning that we are currently unable to
read user for input if we want for instance to define type alias
when sending code action to the server. From the discussion with
`alanz/lsp` devs there're currently 2 ways to workaround with:

> wz1000: custom requests probably - you will have to provide a
> client implementation yourself
> wz1000: Or you could hack it by making users add comments as
> markers in their code
> wz1000: and then make it a code action triggered by the comment
> alanz: awkure, my conclusion is you extract to some made-up name,
> then let the user rename. That is the only thing that allows user
> input at present

So I've chosen the easiest one by allowing the user to rename the
type manually with renaming action.