CI: Build binary without coverage enabled
A user reported that the static binary is spitting out .coverage files
in the working directory.

I presume this is because when we `BISECT_ENABLE=yes dune runtest`,
the executable gets rebuilt with coverage.

When we `LIGO_FORCE_NEW_TYPER=true dune runtest` after this, the
executable does not get rebuilt without coverage, because dune doesn't
know that BISECT_ENABLE matters.

This change works around the problem by making a copy of the
executable after the build, before enabling coverage for tests.