[LIGO-142] Fix let expressions
Problem: In ReasonLIGO, let destructurings don't parse as intended. The
real issue, however, is that ReasonML doesn't understand a sequence of
nested statements in a let-binding.

Solution: Add a let_in production to the parser which understands a
sequence of declarations separated by semi-colon. Add a _program
production which checks whether we have a statement or a final _expr.
Update let_declaration and other relevant fields to make use of _program
which allows the program to be parsed as intended by LIGO. Additionally,
let a program be parsed when it's between parenthesis.