remove par_program and make nonsensial expressions appear in other ones
Problem: I don't like the idea of using `par_program` node since
it's a TS artefact and will shoot in the foot once we start to
use its nested variants (learned the hard way) . And since this
`par_program` node is used only in `_expr` block but `_program`
is not and only remains for getting out `let_in` out of `_expr`,
it confused me at least. Let's be more clear. Since we do not know
how more heretical ligo expressions may be -- let's treat their
expressions that are closed in parenthesis as a separate group
so that we won't drop an error (which we actually should due to
a bad taste in code) in cases such:

```
let f = (lam x => x) + (lam y => y)
```

or something.

Solution: Add `expr_group` node that translates to `Seq` once
parsed in reasonligo dialect and remove any occurence of `par_program`.