Refactor ligo exceptions
Problem: Using sums in exceptions is generally considered as bad
practice. In `AST.Scope.Standard` we have three handlers that do
exactly one thing, return fallback scopes in case that *some*
of the variants of exceptions, defined in sum are failing and
extract meaningful (expected) error message in others. Same goes
to all other `SomeException`s.

Solution: Make every ligo exception a separate type, so that we
can handle them separately (internally all exceptions are treated
as sum variants, so that's not a problem) so that we can handle
*some* particular exceptions and use the default value for
*all the other ones* that are stored in `SomeException` (as the
rest of ligo exceptions as well). You can see the benefits of such
approach in `AST.Scope.Standard` module diff.