Add `mergeScopeForests` function
Problem: We want to be able to merge ligo and fallback scopes for
extracting *full* scope information in a contract.

Solution: Add `mergeScopeForests` function that merges two scopes
in favour of one of them finding ligo ones. Meaning that it replaces
one of them if they are produced from ligo binary, otherwise it
could be thought of just "taking the union" of two scope trees.