vscode: Avoid complex script path resolution
Our platform-dependent dispatcher script used a bit flaky logic for
determining the location of platform-specific binaries to dispatch too.
It start off using `BASH_SOURCE`, which was bash-specific, and then
switched to `readlink`, which caused hard-to-track-down issues.

Avoid any external commands (such as `readlink` and `dirname`) and
instead leverage the full power of a truly Turying-complete language
(TypeScript, that is) to set the working directory of our dispatcher
script to the right one. This allows all paths to be simply relative.