[LIGO-158] Fix find capability when a variable name is a wildcard
Problem: When we have a definition such as `let _ = x * x`, the LSP
throws a `TreeDoesNotContainName` error which prevents capabilities such
as find from working.

Solution: Add a workaround: if we found some wildcard pattern, rebuild
it as a `NameDecl "_"`. Also add a `Range` for `TreeDoesNotContainName`
so we can better identity where an error occured.