[LIGO-164] Fix comments on newline
Problem: When we have a comment in the last line of a file and such line
is non-newline-terminated, such comment will fail to parse.

Solution: Make comment rules not care about newline. Fix a problem in
the parser where the last character would always be dropped; instead if
we find a newline, replace it with nothing.