[LIGO-147] Add module access
Problem: Although we can parse grammars containing `Foo.bar`, we don't
really produce the correct production as `Foo` is parsed as a
`ConstrName`. Also, we can't parse nested modules such as `Foo.Bar.baz`.

Solution: Add two productions: one for accessing a typename in a module
and another for accessing a name in a module. ReasonLIGO needed some
rewriting for spreads to avoid conflicts, and so allow nested fields
only in the case of spreads.

Signed-off-by: Heitor Toledo Lassarote de Paula <heitor.toledo@serokell.io>