[LIGO-159] Report more errors
Problem: The LSP only reports parsing errors, skipping other information
provided by the LIGO compiler.

Solution: These are basically the steps needed:
1. Sometimes a range is printed as natural language rather than a
JSON-ended LigoRange. In case we get a String, use a Regex to extract
the relevant information (file, line, column start, column end);
2. Sometimes the LIGO compiler crashes. Sadly, there is not much we can
do about it, and so we use a Regex to extract the crash reason and
report it.

Additionally, some tests were added, which I hope that are not too
volatile since they rely on the LIGO compiler.