[LIGO-116] Add production for extra parenthesis
Problem: We should add a production for parenthesis and some tests
checking it works.

Solution: Add Paren to the Haskell part of the parser and add or reuse
existing rules in the grammars to accept a parenthesis. Make a
dinstinction from a tuple by having a tuple expect at least 2 parameters
(separated by comma).