[LIGO-168] Fix comments inside block comments
Problem: When we have a block comment with a line comment inside, such
as `(* a comment with // inside of it *)`, the parser trips up.

Solution: Just changing `grammar.js` apparently isn't enough to fix it.
For each grammar, create a `src/scanner.c` file which acts as a external
scanner, which detects comments and generates the appropriate tokens for
each of them. Change the grammar itself so we don't use our RegExes
anymore and add the new tokens as externals. Some files such as
`binding.gyp` and the external scanners would be ignored by Git, so
change the `.gitignore` file a bit so that these two are not ignored.

Please see: https://github.com/tree-sitter/tree-sitter-rust/issues/6