[LIGO-101] Fix expressions longer than they should be
Problem: Errors are reported as one expression longer than they actually
should be. For example, `let _ = * x` might be reported as incorrect in
its entirety, while it should report only the `*` as incorrect.

Solution: Return `Range` together with the error instead of using the
one provided in `before`. Do some gardening on the way while debugging
the issue.