[LIGO-116] Check for empty nodes
Problem: When we have a missing token, tree-sitter won't report an ERROR
node but instead will create an empty node.

Solution: We check if such node exists, and if so, rename it so it's an
ERROR node. Note tree-sitter has some functions to report whether an
error exists, but it seems the Haskell version doesn't really bridge
such function. Please see:
https://github.com/tree-sitter/tree-sitter-bash/issues/27#issuecomment-410865045