[LIGO-172] Split test suite
Problem: We want the `Parsers` tests for upstream contracts separated
from the `Parsers` tests for the LSP contracts.

Solution: First we add all the tests to a new `lsp-test` subdirectory.
After this, we duplicate the `Parsers` test in its own directory and
change both so that redundancies are removed. Then we create a
`test-commons` directory and move common parts of both test suites to
it.