[LIGO-104] Handle included files
Problem: So far, we ignored #include directives. It is desirable that
our capabilities work on included files. Notice we need to check both
for included files and files that include the current one.

Solution: There is a lot in the MR, but basically we can summarize it as
follows:

* Change our parsers so they can scan some directory for contracts. Have
them build an "inclusion graph" by analyzing each file for includes.
* Change scope forests in the following manner:
  * Fallback: DFS the contracts, adding scopes as normally (top-down).
  After the recursion returns, mix the child scopes into the parent
  scopes. This allows us to find references in child files from the
  parent files.
  * FromCompiler: Just dumbily call the LIGO compiler on each dependency
  and do nothing special. Although this has O(n²) complexity. :/
* addLocalScopes: Calculate the union of all scopes into a single scope
forest then, for each scope forest, take the intersection of this
monolithic scope forest and the current forest. This allows us to solve
the case where some contract A includes C and B includes C. Just with
the scope forests alone, A would have the scopes from C, and B would
have the scopes from C, as well as references, but B wouldn't have the
references from C and vice-versa, and neither C would have the
references from A and B. In other words, if C declares some c variable,
it will contain its references in both A and B, for instance.
* mergeScopeForest: Fixed a bug with it that was overlooked on LIGO-99.
If we have some scopes A, B and C on the left side and C, A, B on the
right side, they would likely not even merge. So now we have a smarter
algorithm for trying to find matching scopes so they can be merged.
Additionally, this function now can merge scopes following some
strategy: that is, only merge if both scopes exist (intersection) or
merge normally (union).
* Make RIO scan for the root directory (if present, otherwise just
extract current contract), so it takes a lot of contracts into
consideration.
* Change Rename capability a bit so that it can correctly rename files
in multiple files. Add a new test for this as well.

This MR, however, has some limitations, and follow up issues should be
created. Notably:
* We assume includes are always top level. This may not work as expected
with includes in unorthodox places.
* We should have a mechanism to automatically detect changed documents
and incrementally change the graph on demand. This would eliminate the
need for reparsing contracts every time one of them changes, but scopes
would need to be recalculated from scratch every time.
* Make FromCompiler scopes not fail for everything if one of them fails.
* Make FromCompiler scopes faster. Maybe at each level call the LIGO
compiler asynchronously.
* Additionally, we should work on LIGO-148 soonish so that we can
improve some lag that was introduced here. Otherwise, just investigate
some way to cancel previous requests on timeout.