[LIGO-99] Add some utilities
Problem: We need some utilities which will be useful later on for
LIGO-99. Notably, some way to quickly `nub` things from a list and some
way to `zipWith` two lists that don't discard elements.

Solution: Add `nubOnOrd` (and `nubOrd`) which can nub a list in O(n log
n) (in comparison to the original `nub` which is O(n²). Add
`zipWithRepeat` which duplicates elements at the end of a list.