[LIGO-99] Merge scopes
Problem: Our scopes are not properly merged in some situation. Notably,
if two ranges are disjoint, we may get an error from `Duplo` that
subtrees are not less of [sic] equal. Additionally, none of the scopes
are perfect, and sometimes a definition may be missing from one or the
other, in which such case we don't want to lose them with 'zipWith',
hence the definition of 'zipWithRepeat' before. Additionally, sometimes
there are infinite loops in the `go` function.

Solution: For the union of declarations, we want to join two
`ScopedDecl`s by merging its references and docs, preferring our
fallback's declarations. For the union of scopes, we continue with a
logic that is similar to before, but now we don't merge disjoint ranges,
creating two new decls instead. When we merge `DeclRef`s, we prefer the
one with smallest scope, as it is more local and improves UX. E.g.: we
prefer `n` rather than `n : int`. To fix the infinite loops in `go`, we
choose the immediate "parent" of some `ScopeTree` instead of the root
`ScopeTree` in the case they are not equal.

Finally, test that both our Fallback and Standard parses work. If
Fallback works, then Standard should also work since it should include
everything that Fallback contains.