[LIGO-160] Disable let-destructuring tests
Problem: The compiler added functionality for deep pattern matching,
which is not present in the language server currently. This is blocking
merging with the upstream, and will take a significant amount of effort
to implement.

Solution: Disable the deep pattern matching tests for now, and create
and issue to implement them later.