[LIGO-179] Fix unresolved type parser
Problem: Our FromCompiler scopes frequently fail because the compiler
may return an unresolved type which we don't parse.

Solution: This was copied from LIGO-159's !122 MR, which we put some
value inside a LigoTypeFull in case of unresolved. I'm not sure what
kind of values may go in, but I haven't seen anything different than
null so this will do for now.