[LIGO-130] Bump tree-sitter version
Problem: We use a somewhat old version of tree-sitter (0.16.9) so it
would be desirable to use a more recent version (0.19.5).

Solution: For now, we are using our own fork with updated tree-sitter.
We fetch it in stack.yaml and cabal.project. The directory structure
changed a bit, so we also updated the .gitignore to filter unwanted new
files. Change Nix a bit so it uses the correct version of the
tree-sitter binary as well.