[LIGO-206] Bump hackage and stackage
Problem: hackage.nix and stackage.nix we pin don't know about LTS 18.0.

Solution: Bump both with `nix flake lock --update-input hackage-nix
--update-input stackage-nix`. Use a newer revision of `haskell.nix`.
Some errors arose from this, notably:
* Nix doesn't recognize `.source` in some places anymore. For
`haskell-nix.sources` I just removed this part. Later on, I replaced
`haskell-nix` with `haskellNix`.
* `package.ghcOptions` was not recognized anymore. I've used
`ghcOptions` which according to the docs seem to be valid, and it
complained that it expects a list of strings, so add a list around it as
well.