[LIGO-211] Fix contracts not accepted by compiler
Problem: Some contracts in `tools/lsp/squirrel/test/contracts` are not
accepted by the LIGO compiler. Some contracts are meant to fail, but
some should not fail.

Solution: Fix the contracts that are not supposed to fail on the
compiler. Leave the contracts that are designed to fail in place.