[LIGO-97] Move graph algorithms to its module, add WCC
Problem: Before we can close diagnostics, we need to first filter
diagnostics by their graph group. This is because even if we open some
contract, it will add error diagnostics for all other contracts, even
the ones that are completely unrelated to it. It is desirable that first
we filter which contracts are related to which ones first, and so first
we need to create an algorithm that does such a thing before we can
begin refactoring.

Solution: First, create a new `Util.Graph` module that will have
graph-related things, as they shouldn't be in `AST.Scope.Common`.
Created a function for computing the WCC (Weakly Connected Components)
of a graph, which will return all subgraphs such that all vertices are
mutually reachable from one another.