[LIGO-97] Handle diagnostics
Problem: Our diagnostics may persist open even if a document is closed.
We want to close them whenever a document is closed.

Solution: It's a bit subjective, but I thought it would be best to keep
all errors related to some contract while at least one of them is
related. That is, if there is some group of files whose includes are
part of a WCC, errors for them will be cleared iff all of them are
closed. When one of them opens or gets edited, the others may have
errors as well.