[LIGO-216] Use blacklist instead of whitelist
Problem: When we test the behavior of the extension against the contracts in
src/test/contracts, we use a whitelist to decide which directories to recurse
into. This is problematic, since we will miss new directories that the client
adds.

Solution: Use a blacklist instead of a whitelist to decide which
subdirectories of src/test/contracts to consider.