[LIGO-185] Use LIGO preprocess
Problem: We should have a function that will call `ligo preprocess` for
us and use it when parsing contracts.

Solution: First refactor `getLigoDefinitionsFrom` so that its error
handling part is a separate function, `handleLigoError`. Write
`preprocess` to call `ligo preprocess` and create a new exception
indicating that the preprocessor failed.

I noticed that there is a bit of inconsistency on how LIGO may report
locations in errors, so the Regex had to be changed a bit.

Change our parsers so that they are able to extract the ranges. In
particulary, `extractIncludedFiles` had to be radically changed, as well
as `includesGraph`. Since `extractIncludedFiles` now contains the core
for dealing with included files, this is how to interpret the algorithm:

We first traverse the tree top-down, collecting all line markers we find
along the way and storing some metadata for processing. Simply put, we
want to map the preprocessed tree back to what it was when there were no
line markers, while keeping all expressions visible. We remap the range
so that included files are "compressed" into the line where the first
include would be. Meanwhile, elements from the "root" tree are just
mapped to whatever the line markers indicate. This changes the starting
line of everything. After we traversed the entire tree, we begin
rebuilding it top-down, where we now change the finishing range of
everything (as we may see new line markers that we didn't visit yet
during top-down phase). We also add a preprocessed range that maps to
the "real" locations, including the preprocessed files.