[LIGO-230] Prompt user when there is a new LIGO version
Problem: We want the user to be prompted when a new LIGO version is
available.

Solution: Check the installed LIGO. If it's not available, ask the user
to provide a path to it or to download it. Otherwise, check from the
GitLab API if the newest release is better than the current release, and
if so, open the releases page so that the user can download it.