[LIGO-230] Update LIGO for static binary
Problem: We can update LIGO in case of a static binary, so why not do
it?

Solution: If the user has chosen to update automatically, then use the
provided LIGO path to download. We use VSCode's and Axios' capabilities
to support cancellation and progress reporting. At the end, overwrite
the file with 755 permission mode.