[LIGO-196] Store version number in open contracts
Problem: There isn't much about contract versions in VSCode, but from my
investigation and looking at how other extensions do it, it seems they
are only stored for open contracts (if some contract is on the disk but
not open, we use nothing to indicate the disk is master). Anyway, we
wish to store version numbers.

Solution: Following the steps of the C++ extension for VSCode, store the
versions of the currently open documents, and forget about the ones that
were closed.