[LIGO-150] Extract global config
Problem: The VSCode client was not sending the configuration to the
server.

Solution: It turns out that now the LSP server has to manually request
the configuration through the `workspace/configuration` request. We add
functionality to ligo-squirrel to request the configuration upon
initialization. Unfortunately, the Haskell lsp library doesn't provide a
way to update the Config in the LspM monad, so we maintain our own
Config in the RIO monad.