[LIGO-149] Prettify information on hover
Problem: Currently, when we try to hover over some type or variable, we
get a mixture of our s-exps pretty-printers and dialect-specific
pretty-printers. We wish to make it a bit more pretty.

Solution: Parse some more types to a "typed" representation. Improve the
pretty-printing for type application for all dialects. Improve some
hover specific information a bit more.