[LIGO-240] Fix mapConcurrently
Problem: mapConcurrently spawns a thread for each element in the
collection. This is a problem for any list that is quite small, as it
will use a lot of resources to handle all threads.

Solution: Divide the list into bins according to the number of
concurrent capabilities the computer has.