[LIGO-178] Use better threaded flags
Problem: We should ensure that our flags are optimal for the
application.

Solution: I tried many flags configurations but didn't really see much
difference between them. Ultimately, I think it's difficult to really
benchmark the application with the current performance issues we have,
and so I decided to only change the idle GC timeout and add threaded
runtime flags to the executables.