[LIGO-249] Fix extension not working without Internet connection
Problem: If the user opens the VSCode extension without an Internet
connection, then the extension won't activate.

Solution: There was actually only one problem: in
`getCurrentExtensionVersion`, we should check if the `data` field was
undefined/null or not. I did some other cleanup that hopefully makes
the code a bit saner.
1. We have an `ExtensionQuery` type instead of just using `any` blindly.
2. We use `if (!foo)` instead of `if (foo === undefined)`.
3. When catching an error and showing it through VSCode, we return
   `undefined` instead of whatever VSCode's windows return, which I
   believe is whatever buttons are avaialble.