[LIGO-248] Fix 'Choose path' on VSCode
Problem: We don't see the path to the ligo binary until after the
extension is restarted, if the user has used the 'Choose path' button.

Solution: Simply fetch the configuration again and use it. You'd expect
that `update` would actually update the reference, but it doesn't, you
need to get the configuration again.