[LIGO-250] Handle configuration changes
Problem: We can extract the configuration at startup, but not when it
changes while the extension is running.

Solution: Register that we can change configuration. Handle
WorkspaceDidChangeConfiguration and update the configuration. Note that
for some reason, I always have to fetch the configuration (that is, in
`updateCustomConfig`, we end up having to call `fetchCustomConfig` all
the time). At least it's better than the old approach of always calling
it, since now we only fetch when it indeed changed. This happens because
we seem to always get `null` for some reason...

On the way, did some gardening to use the unlifted versions of the
`MVar` functions.